feat: añado endpoints, DTOs, tarifa (service y controller)

+++ ms-camiones/src/main/java/com/backend_tpi/ms_camiones/models/Tarifa.java
@@ -20,5 +20,5 @@ public class Tarifa {
     private String concepto;
 
     @Column(name = "COSTO_X_KILOMETRO", nullable = false)
-    private BigDecimal costoXKilometro;
+    private Double costoXKilometro;
 }
